+++ CHANGELOG.md
@@ -17,6 +17,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Initial support for outputting a single value using a `min()` function, rather than a media query (#79)
 - The `max` option can now be set to a `var()` function (#79)
 - Properties are deprecated and will be removed in a future version: `tidy-column`, `tidy-offset`, `tidy-span`, `tidy-offset-left`, `tidy-offset-right` (#81)
+- `debug` will now print a warning, rather than preserve the declaration as a comment
 
 **Removed**